context and react query add to searchView
